requirements.txt
spacy
textblob
pdf2image 
pytesseract
Pillow
pandas
wordcloud
matplotlib
streamlit
fastapi